refactor: fix get_logger import paths
